fix: fix bugs with repeated option

+++ test/php/src/cyamli/Input_Version.php
@@ -83,7 +83,7 @@ class Input_Version {
             case 'int[]':
                 $val = [];
                 foreach ($strValue as $str) {
-                    $v = self::parseValue('int64', $str);
+                    $v = self::parseValue('int', $str);
                     if ($v === null) return null;
                     $val[] = $v;
                 }
