+++ test/php/src/features/Input_Sub1Sub2.php
@@ -83,7 +83,7 @@ class Input_Sub1Sub2 {
             case 'int[]':
                 $val = [];
                 foreach ($strValue as $str) {
-                    $v = self::parseValue('int64', $str);
+                    $v = self::parseValue('int', $str);
                     if ($v === null) return null;
                     $val[] = $v;
                 }
